Backend(Cloud Engine)/requirements.txt
# Web Framework / Server
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==15.0.1
starlette==0.47.3

# AI / Machine Learning Libraries
funasr==1.2.7
fer==22.5.1
tensorflow==2.20.0
opencv-python==4.12.0.88
numpy==2.3.3
soundfile==0.13.1
moviepy==1.0.3
imageio==2.37.0
imageio-ffmpeg==0.6.0

# Google Cloud Libraries
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.29.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
requests==2.32.5

# Firebase
firebase-admin==7.1.0
google-cloud-firestore==2.21.0

# Data Validation & Processing
pydantic==2.11.9

# Utilities
python-dateutil==2.9.0.post0
pillow==11.3.0